Description translation suggestion endpoint
Adds an endpoint for suggesting Wikipedia articles in a target language
for which a description is needed for the corresponding Wikidata entity,
given that an article referring to the same Wikidata entity exists in
another language and also has a description in that language.

Recommendation generation follows the same basic pattern as with the
article translation suggestion endpoint: a pool of candidates is generated
(here, from the MW action API using action=query&generator=random); and
results are filtered first according to their own properties, and then
with the results of a follow-up Wikidata Query Service query.

When targeting English, an additional filtering query is performed via the
enwiki action API in order to filter out candidates with local descriptions
but not Wikidata descriptions.